third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/custom-element-lifecycle-events.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/console-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script>

function test() {
  var setup = [
    'var proto = Object.create(HTMLElement.prototype);',
    'proto.createdCallback = function() { output(\'Invoked createdCallback.\'); };',
    'proto.attachedCallback = function() { output(\'Invoked attachedCallback.\'); };',
    'proto.detachedCallback = function() { output(\'Invoked detachedCallback.\'); };',
    'proto.attributeChangedCallback = function() { output(\'Invoked attributeChangedCallback.\'); };',
    'CustomElement = document.registerElement(\'x-foo\', {prototype: proto});',
  ].join('\n');

  var lifecycleCallbacks = [
    'created = new CustomElement();',
    'created.setAttribute(\'x\', \'1\');',
    'document.body.appendChild(created);',
    'created.remove();',
  ].join('\n');

  SourcesTestRunner.startDebuggerTest(step1);

  function step1() {
    ConsoleTestRunner.evaluateInConsole(setup, function() {
      TestRunner.addResult('Custom element registered.');
      ConsoleTestRunner.evaluateInConsoleAndDump('new CustomElement() instanceof CustomElement', step2);
    });
  }

  function step2() {
    ConsoleTestRunner.evaluateInConsole('debugger;');
    SourcesTestRunner.waitUntilPaused(step3);
  }

  function step3() {
    ConsoleTestRunner.evaluateInConsoleAndDump('new CustomElement() instanceof CustomElement', step4);
  }

  function step4() {
    ConsoleTestRunner.evaluateInConsole(lifecycleCallbacks, SourcesTestRunner.completeDebuggerTest);
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that custom element lifecycle events fire while debugger is paused.
</p>

</body>
</html>
