third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/custom-element-lifecycle-events-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that custom element lifecycle events fire while debugger is paused.

Custom element registered.
new CustomElement() instanceof CustomElement = true
Script execution paused.
new CustomElement() instanceof CustomElement = true
Script execution resumed.
[page] Invoked createdCallback.
[page] Invoked createdCallback.
[page] Invoked createdCallback.
[page] Invoked attributeChangedCallback.
[page] Invoked attachedCallback.
[page] Invoked detachedCallback.
